feat: upgrade MetaMask SDK to 2.1.0
