+++ Assets/Thirdweb/Core/Plugins/MetaMask/Installer/Documentation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cdb342bc09fb445a2b558fa4d120410e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
