+++ Assets/Thirdweb/Core/Plugins/MetaMask/Scripts/Providers/Sockets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bde9e40c622dbbd4887d5f1ec906472e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
