--- Assets/Thirdweb/Core/Plugins/MetaMask/Scripts/Sockets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bde9e40c622dbbd4887d5f1ec906472e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
