--- Assets/Thirdweb/Core/Plugins/MetaMask/Scripts/Sockets/MetaMaskUnitySocketIO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c9b8fb604f0d63478d1c1298c371be2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
